change ordering of json object being created
json object is now passed in as a PSCustomObject, instead of being created for each function. This should improve speed, but also allows us to test the json file before we even make a connection to the database. Test-Json can also be used by devs to validate json file before checking in.